pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-core>=0.3.55",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.0",
    "reportlab>=4.4.0",
    "streamlit>=1.44.1",
]
